separate gpu and dependencies
